feat: add vue-route logic
