+++ frontend/package.json
@@ -19,6 +19,7 @@
   },
   "dependencies": {
     "axios": "^1.7.7",
-    "vue": "^3.5.5"
+    "vue": "^3.5.5",
+    "vue-router": "^4.4.5"
   }
 }
