+++ frontend/src/view/AboutView.vue
@@ -1,0 +1,3 @@
+<template>
+	<div>about view</div>
+</template>
