+++ frontend/vite.config.js
@@ -1,8 +1,7 @@
-import {defineConfig} from "vite";
+import { defineConfig } from "vite";
 import vue from "@vitejs/plugin-vue";
 
-
 export default defineConfig({
-    base: '/todo-app',
-    plugins: [vue()]
+	base: "/todo-app",
+	plugins: [vue()],
 });
